requirements.txt
python-dotenv

mcp[cli]
strands-agents
strands-agents-tools

streamlit
boto3 
vecs

supabase
tavily-python